requirements.txt
kubernetes==11.0.0
pyyaml==5.3.1
celery==4.4.6
redis==3.5.3